+++ reactivestate/src/main/java/com/ensody/reactivestate/android/ViewModelExt.kt
@@ -6,6 +6,64 @@ import androidx.fragment.app.Fragment
 import androidx.fragment.app.activityViewModels
 import androidx.fragment.app.viewModels
 import androidx.lifecycle.*
+import com.ensody.reactivestate.NamespacedStateFlowStore
+import com.ensody.reactivestate.StateFlowStore
+import kotlinx.coroutines.CoroutineScope
+import kotlin.reflect.KClass
+
+/**
+ * Creates an object living on a wrapper `ViewModel`. This allows for building multi-platform ViewModels.
+ *
+ * The [provider] should instantiate the object directly.
+ *
+ * @see [buildViewModel] and [stateViewModel] if you want to instantiate an Android `ViewModel` directly.
+ */
+public inline fun <reified T : Any> Fragment.buildOnViewModel(
+    crossinline provider: BuildOnViewModelContext.() -> T,
+): Lazy<T> =
+    stateViewModel { WrapperViewModel(it) }.buildOnViewModel(provider)
+
+/**
+ * Creates an object living on a wrapper `ViewModel`. This allows for building multi-platform ViewModels.
+ *
+ * The [provider] should instantiate the object directly.
+ *
+ * @see [buildViewModel] and [stateViewModel] if you want to instantiate an Android `ViewModel` directly.
+ */
+public inline fun <reified T : Any> ComponentActivity.buildOnViewModel(
+    crossinline provider: BuildOnViewModelContext.() -> T,
+): Lazy<T> =
+    stateViewModel { WrapperViewModel(it) }.buildOnViewModel(provider)
+
+/** Build context for [buildOnViewModel]. */
+public class BuildOnViewModelContext(
+    /** The [ViewModel.viewModelScope]. */
+    public val scope: CoroutineScope,
+
+    /** A [StateFlowStore] where you can store/load the saved instance state (similar to a [SavedStateHandle]). */
+    public val stateFlowStore: StateFlowStore,
+)
+
+/** The wrapper ViewModel used by [buildOnViewModel]. */
+public class WrapperViewModel(savedStateHandle: SavedStateHandle) : ViewModel() {
+    public val stateFlowStore: StateFlowStore = SavedStateHandleStore(viewModelScope, savedStateHandle)
+    public val registry: MutableMap<KClass<*>, Any> = mutableMapOf()
+}
+
+/** Used internally by [buildOnViewModel]. */
+public inline fun <reified T : Any> Lazy<WrapperViewModel>.buildOnViewModel(
+    crossinline provider: BuildOnViewModelContext.() -> T,
+): Lazy<T> = lazy {
+    val stateFlowStore = NamespacedStateFlowStore(
+        store = value.stateFlowStore,
+        namespace = T::class.qualifiedName
+            ?: throw IllegalArgumentException("The class must have a qualifiedName"),
+    )
+    val result = value.registry[T::class] as? T
+        ?: BuildOnViewModelContext(scope = value.viewModelScope, stateFlowStore = stateFlowStore).provider()
+    value.registry[T::class] = result
+    result
+}
 
 /**
  * Creates `ViewModel`.
@@ -14,7 +72,7 @@ import androidx.lifecycle.*
  */
 @Suppress("UNCHECKED_CAST")
 public inline fun <reified T : ViewModel> Fragment.buildViewModel(crossinline provider: () -> T): Lazy<T> =
-    viewModels<T> {
+    viewModels {
         object : ViewModelProvider.Factory {
             override fun <T : ViewModel?> create(modelClass: Class<T>): T = provider() as T
         }
@@ -27,7 +85,7 @@ public inline fun <reified T : ViewModel> Fragment.buildViewModel(crossinline pr
  */
 @Suppress("UNCHECKED_CAST")
 public inline fun <reified T : ViewModel> ComponentActivity.buildViewModel(crossinline provider: () -> T): Lazy<T> =
-    viewModels<T> {
+    viewModels {
         object : ViewModelProvider.Factory {
             override fun <T : ViewModel?> create(modelClass: Class<T>): T = provider() as T
         }
@@ -40,7 +98,7 @@ public inline fun <reified T : ViewModel> ComponentActivity.buildViewModel(cross
  */
 @Suppress("UNCHECKED_CAST")
 public inline fun <reified T : ViewModel> Fragment.activityViewModel(crossinline provider: () -> T): Lazy<T> =
-    activityViewModels<T> {
+    activityViewModels {
         object : ViewModelProvider.Factory {
             override fun <T : ViewModel?> create(modelClass: Class<T>): T = provider() as T
         }
@@ -55,7 +113,7 @@ public inline fun <reified T : ViewModel> Fragment.activityViewModel(crossinline
 public inline fun <reified T : ViewModel> Fragment.stateViewModel(
     crossinline provider: (handle: SavedStateHandle) -> T,
 ): Lazy<T> =
-    viewModels<T> {
+    viewModels {
         object : AbstractSavedStateViewModelFactory(this, null) {
             override fun <T : ViewModel?> create(
                 key: String,
@@ -74,7 +132,7 @@ public inline fun <reified T : ViewModel> Fragment.stateViewModel(
 public inline fun <reified T : ViewModel> ComponentActivity.stateViewModel(
     crossinline provider: (handle: SavedStateHandle) -> T,
 ): Lazy<T> =
-    viewModels<T> {
+    viewModels {
         object : AbstractSavedStateViewModelFactory(this, null) {
             override fun <T : ViewModel?> create(
                 key: String,
@@ -93,7 +151,7 @@ public inline fun <reified T : ViewModel> ComponentActivity.stateViewModel(
 public inline fun <reified T : ViewModel> Fragment.activityStateViewModel(
     crossinline provider: (handle: SavedStateHandle) -> T,
 ): Lazy<T> =
-    activityViewModels<T> {
+    activityViewModels {
         object : AbstractSavedStateViewModelFactory(requireActivity(), null) {
             override fun <T : ViewModel?> create(
                 key: String,