+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## Next release
+
+* Added `buildOnViewModel` which allows creating arbitrary objects living on an internally-created wrapper ViewModel.
+* Added `MutableStateFlow.replace` helper function for simplifying e.g. data class `copy()` based mutation.
+
 ## 3.3.0
 
 * Added `NamespacedStateFlowStore`.